move eservice not found from models to catalog process
