+++ packages/catalog-process/src/services/catalogService.ts
@@ -17,7 +17,6 @@ import {
   WithMetadata,
   catalogEventToBinaryData,
   descriptorState,
-  eServiceNotFound,
 } from "pagopa-interop-models";
 import { match } from "ts-pattern";
 import { v4 as uuidv4 } from "uuid";
@@ -57,6 +56,7 @@ import {
   eServiceDuplicate,
   notValidDescriptor,
   operationForbidden,
+  eServiceNotFound,
 } from "../model/domain/errors.js";
 import { readModelService } from "./readModelService.js";
